+++ Retrospective/Core/Protocols/KeyboardReadable.swift
@@ -1,11 +1,10 @@
 //
-//  KeyboardStatus.swift
+//  KeyboardReadable.swift
 //  Retrospective
 //
 //  Created by 정소이 on 5/15/25.
 //
 
-import Foundation
 import SwiftUI
 import Combine
 
@@ -25,29 +24,5 @@ extension KeyboardReadable {
                 .map { _ in false }
         )
         .eraseToAnyPublisher()
-    }
-}
-
-final class ContentViewModel: ObservableObject, KeyboardReadable {
-    enum Action {
-        case shouldResignKeyboard
-    }
-
-    /// 키보드가 열려있는지 여부
-    @Published var isShowingKeyboard: Bool = false
-
-    init() {
-        bind()
-    }
-
-    private var cancellables = Set<AnyCancellable>()
-
-    private func bind() {
-        keyboardPublisher
-            .removeDuplicates()
-            .sink { [weak self] isShowing in
-                guard let self else { return }
-                self.isShowingKeyboard = isShowing
-            }.store(in: &cancellables)
     }
 }
